+++ src/ui/page/PageTitle.vue
@@ -1,6 +1,6 @@
 <template>
   <q-toolbar>
-    <q-item-label header class="q-pl-none text-h6">
+    <q-item-label header class="q-pl-none text-h6 text-uppercase">
       {{ title }}
     </q-item-label>