+++ src/markdowns/changelog.md
@@ -1,3 +1,11 @@
+Version: 0.1.2
+Date: June 16, 2020
+```
+Features:
+- Fixed project menu to show only relevant menu items
+- New: Validate Project page to facilitate validate/return of endorsed project
+```
+
 Version: 0.1.1
 Date: June 8, 2020
 ```